atualizacao

+++ previsao_trens/packages/PROG_SUBIDA/CALCULAR_SUBIDA_V2.py
@@ -1171,7 +1171,7 @@ class Condensados():
         from previsao_trens.models  import TremVazio
         from django.db.models       import Q, Sum, F
         
-        
+
         dict_segmentos = { "qt_graos": "GRAO", "qt_ferti": "FERTILIZANTE", "qt_celul": "CELULOSE", "qt_acuca": "ACUCAR", "qt_contei": "CONTEINER" }
         
         previsao    = (dict_trem["previsao"]  - timedelta(hours=1))
@@ -1179,24 +1179,26 @@ class Condensados():
         hora_fim    = hora_inicio + timedelta(hours=1)
         
         trens_da_previsao = TremVazio.objects.filter(Q(previsao__gte=hora_inicio) & Q(previsao__lt=hora_fim))
-           
-        data_arq = (previsao).strftime('%Y-%m-%d')
+        
+        data_arq = (previsao).strftime('%Y-%m-%d')   
+        path_condensado = f"previsao_trens/src/SUBIDA/CONDENSADOS/condensado_{ data_arq }.json"
+        
+        if os.path.exists(path_condensado):
 
-        with open(f"previsao_trens/src/SUBIDA/CONDENSADOS/condensado_{ data_arq }.json") as arquivo_json:
-            consensados = json.load(arquivo_json)
+            with open(path_condensado) as arquivo_json:
+                consensados = json.load(arquivo_json)
 
-        for ferrovia in ["RUMO", "MRS", "VLI"]:
+            for ferrovia in ["RUMO", "MRS", "VLI"]:
 
-            trens_da_ferrovia = trens_da_previsao.filter(ferrovia=ferrovia)
+                trens_da_ferrovia = trens_da_previsao.filter(ferrovia=ferrovia)
 
-            for segmento in dict_segmentos:
-                
-                total_segmento = trens_da_ferrovia.aggregate(total=Sum(F(segmento)))['total']
-                if total_segmento == None: total_segmento = 0
-                consensados[dict_trem["margem"].upper()]["SAIDAS"]["FERROVIA"][previsao.hour][ferrovia][dict_segmentos[segmento]] = total_segmento
-
+                for segmento in dict_segmentos:
+                    
+                    total_segmento = trens_da_ferrovia.aggregate(total=Sum(F(segmento)))['total']
+                    if total_segmento == None: total_segmento = 0
+                    consensados[dict_trem["margem"].upper()]["SAIDAS"]["FERROVIA"][previsao.hour][ferrovia][dict_segmentos[segmento]] = total_segmento
             
-        with open(f"previsao_trens/src/SUBIDA/CONDENSADOS/condensado_{ data_arq }.json", 'w') as arquivo_json:
-            json.dump(consensados, arquivo_json, indent=4)
+            with open(path_condensado, 'w') as arquivo_json:
+                json.dump(consensados, arquivo_json, indent=4)
 
-        SUBIDA_DE_VAZIOS().ATUALIZAR()
+            SUBIDA_DE_VAZIOS().ATUALIZAR()
